# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alter Event Tracker",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register a new staff account (pending admin approval)",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffRegisterRequest"
              },
              "examples": {
                "new_volunteer": {
                  "summary": "New volunteer registration",
                  "description": "Register a new staff account. Account will be pending admin approval.",
                  "value": {
                    "username": "nuevo_voluntario",
                    "password": "contrase\u00f1a_segura_123!"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created (pending approval)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                },
                "examples": {
                  "success": {
                    "summary": "Registration successful",
                    "value": {
                      "id": 3,
                      "username": "nuevo_voluntario",
                      "is_approved": false,
                      "role": {
                        "id": 1,
                        "name": "USER"
                      },
                      "created_at": "2024-03-22T21:15:00Z",
                      "is_disabled": false
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Username already taken",
            "content": {
              "application/json": {
                "examples": {
                  "username_taken": {
                    "summary": "Username already registered",
                    "value": {
                      "detail": "Username 'maria_admin' is already taken."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "examples": {
                  "password_too_short": {
                    "summary": "Password too short",
                    "value": {
                      "detail": [
                        {
                          "type": "string_too_short",
                          "loc": [
                            "body",
                            "password"
                          ],
                          "msg": "String should have at least 15 characters",
                          "input": "short",
                          "ctx": {
                            "min_length": 15
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login - returns a JWT bearer token",
        "operationId": "login_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_auth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                },
                "examples": {
                  "success": {
                    "summary": "JWT token returned",
                    "value": {
                      "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJtYXJpYV9hZG1pbiIsInJvbGUiOiJTSVRFX0FETUlOIiwiZXhwIjoxNzExMjM0NTY3fQ.fake_signature_for_documentation_purposes_only",
                      "token_type": "bearer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Wrong username or password",
            "content": {
              "application/json": {
                "examples": {
                  "wrong_password": {
                    "summary": "Wrong username or password",
                    "value": {
                      "detail": "Incorrect username or password."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Account not approved or disabled",
            "content": {
              "application/json": {
                "examples": {
                  "pending_approval": {
                    "summary": "Account pending approval",
                    "value": {
                      "detail": "Your account is pending admin approval. Please try again later."
                    }
                  },
                  "account_disabled": {
                    "summary": "Account disabled",
                    "value": {
                      "detail": "Your account has been disabled."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/userinfo": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Return the currently authenticated user's profile",
        "operationId": "get_me_auth_userinfo_get",
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                },
                "examples": {
                  "admin_user": {
                    "summary": "Admin viewing their profile",
                    "value": {
                      "id": 1,
                      "username": "maria_admin",
                      "is_approved": true,
                      "role": {
                        "id": 2,
                        "name": "SITE_ADMIN"
                      },
                      "created_at": "2024-03-15T20:00:00Z",
                      "is_disabled": false
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "examples": {
                  "missing_token": {
                    "summary": "No token provided",
                    "value": {
                      "detail": "Not authenticated"
                    }
                  },
                  "invalid_token": {
                    "summary": "Invalid or expired JWT",
                    "value": {
                      "detail": "Could not validate credentials"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Account not approved or disabled",
            "content": {
              "application/json": {
                "examples": {
                  "pending_approval": {
                    "summary": "Account pending approval",
                    "value": {
                      "detail": "Your account is pending admin approval. Please try again later."
                    }
                  },
                  "account_disabled": {
                    "summary": "Account disabled",
                    "value": {
                      "detail": "Your account has been disabled."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/auth/verify-recovery-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify a 6-digit recovery code without consuming it",
        "description": "Check that a recovery code is valid and not yet expired or used.\n\nDoes NOT mark the code as used, so closing the tab and retrying still works.\nReturns a generic error on invalid username to avoid user enumeration.",
        "operationId": "verify_recovery_code_auth_verify_recovery_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRecoveryCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Verify Recovery Code Auth Verify Recovery Code Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Reset a user's password using a valid recovery code",
        "description": "Reset the user's password if the provided recovery code is valid.\n\nMarks the code as used after a successful reset so it cannot be reused.\nReturns a generic error on invalid username to avoid user enumeration.",
        "operationId": "reset_password_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_login_auth_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_auth_token_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PossibleRoles": {
        "type": "string",
        "enum": [
          "USER",
          "SITE_ADMIN"
        ],
        "title": "PossibleRoles",
        "description": "Global roles stored in user_roles / users.role_id.\n\nUSER:       Regular approved account. Actual capabilities come from\n            user_organizations.role (per-org STAFF or ADMIN).\nSITE_ADMIN: Superadmin. Bypasses all org membership checks, can create\n            and manage organizations."
      },
...
      "ResetPasswordRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 15,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "code",
          "new_password"
        ],
        "title": "ResetPasswordRequest"
      },
      "ResetPasswordResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "ResetPasswordResponse"
      },
      "RolePublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "$ref": "#/components/schemas/PossibleRoles"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "RolePublic",
        "description": "Public representation of a user role.\n\nUsed when returning user data that includes their role information.\nMaps directly to the Role model via from_attributes."
      },
...
      "StaffRegisterRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 15,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "StaffRegisterRequest",
        "description": "Request body for staff registration.\n\nStrict mode ensures exact type matching - no coercion from strings to ints, etc.\nPassword has a minimum length of 15 characters for security."
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse",
        "description": "Response returned after successful authentication.\n\nContains the JWT access token that the client should include in\nsubsequent requests via the Authorization header."
      },
...
      "UserPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "is_approved": {
            "type": "boolean",
            "title": "Is Approved"
          },
          "role": {
            "$ref": "#/components/schemas/RolePublic"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_disabled": {
            "type": "boolean",
            "title": "Is Disabled",
            "default": false
          },
          "disabled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled At"
          },
          "disabled_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "is_approved",
          "role",
          "created_at"
        ],
        "title": "UserPublic",
        "description": "Public representation of a staff user.\n\nReturned when fetching user details. Includes the nested role object\nfor convenience. Never includes the hashed password."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyRecoveryCodeRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "username",
          "code"
        ],
        "title": "VerifyRecoveryCodeRequest"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  },
...
}